♻ Simplfy attempts to giscedata_facturacio_factura, reintents are not needed
