+++ som_card_payment/models/__init__.py
@@ -5,4 +5,4 @@ from . import res_partner
 from . import res_partner_creditcard
 from . import giscedata_polissa
 from . import account_invoice
-from . import card_payment_attempt
+from . import giscedata_facturacio_factura
