+++ som_card_payment/security/ir.model.access.csv
@@ -1,3 +1,2 @@
 "id","name","model_id:id","group_id:id","perm_read","perm_write","perm_create","perm_unlink"
 "access_res_partner_creditcard_user","res.partner.creditcard","model_res_partner_creditcard","base.group_user",1,1,1,1
-"access_som_card_payment_attempt_user","som.card.payment.attempt","model_som_card_payment_attempt","base.group_user",1,1,1,0
